I split the code into modules, removed the try/except that handles the entire function, and added response_model somewhere, but not everywhere it worked.
