+++ .gitignore
@@ -3,7 +3,6 @@
 *.sqlite3
 
 venv/
-env/
 .venv/
 
 __pycache__/
